refactor(services/gdrive): move raw requests to core (#6088)
