+++ core/src/services/gdrive/backend.rs
@@ -19,12 +19,9 @@ use std::fmt::Debug;
 use std::sync::Arc;
 
 use bytes::Buf;
-use bytes::Bytes;
 use chrono::Utc;
-use http::Request;
 use http::Response;
 use http::StatusCode;
-use serde_json::json;
 
 use super::core::GdriveCore;
 use super::core::GdriveFile;
@@ -132,49 +129,7 @@ impl Access for GdriveBackend {
     }
 
     async fn copy(&self, from: &str, to: &str, _args: OpCopy) -> Result<RpCopy> {
-        let from = build_abs_path(&self.core.root, from);
-
-        let from_file_id = self.core.path_cache.get(&from).await?.ok_or(Error::new(
-            ErrorKind::NotFound,
-            "the file to copy does not exist",
-        ))?;
-
-        let to_name = get_basename(to);
-        let to_path = build_abs_path(&self.core.root, to);
-        let to_parent_id = self
-            .core
-            .path_cache
-            .ensure_dir(get_parent(&to_path))
-            .await?;
-
-        // copy will overwrite `to`, delete it if exist
-        if let Some(id) = self.core.path_cache.get(&to_path).await? {
-            let resp = self.core.gdrive_trash(&id).await?;
-            let status = resp.status();
-            if status != StatusCode::OK {
-                return Err(parse_error(resp));
-            }
-
-            self.core.path_cache.remove(&to_path).await;
-        }
-
-        let url = format!(
-            "https://www.googleapis.com/drive/v3/files/{}/copy",
-            from_file_id
-        );
-
-        let request_body = &json!({
-            "name": to_name,
-            "parents": [to_parent_id],
-        });
-        let body = Buffer::from(Bytes::from(request_body.to_string()));
-
-        let mut req = Request::post(&url)
-            .body(body)
-            .map_err(new_request_build_error)?;
-        self.core.sign(&mut req).await?;
-
-        let resp = self.core.info.http_client().send(req).await?;
+        let resp = self.core.gdrive_copy(from, to).await?;
 
         match resp.status() {
             StatusCode::OK => Ok(RpCopy::default()),
